+++ public/index.php
@@ -1,0 +1,3 @@
+<?php
+
+echo '<h1>Hello!!!</h1>';
